backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.1
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
